requirements.txt
Django>=3.2,<3.3
psycopg2>=2.9,<3.0
flake8>=3.9,<3.10
djangorestframework>=3.13,<3.14
gunicorn>=20.1,<20.2